requirements.txt
# ManualQ Requirements

# Core Dependencies
streamlit==1.28.1
python-dotenv==1.0.0

# PDF Processing
PyPDF2==3.0.1
pymupdf==1.23.8
pdfplumber==0.10.3

# AI & Embeddings
openai==1.3.0
sentence-transformers==2.2.2
langchain==0.0.352
langchain-community==0.0.10

# Vector Search
faiss-cpu==1.7.4
numpy==1.24.3
scipy==1.11.4

# Text Processing
nltk==3.8.1
regex==2023.10.3

# Data Handling
pandas==2.1.2
python-dateutil==2.8.2

# Utilities
requests==2.31.0
pydantic==2.4.2
tqdm==4.66.1
